Module CFG

module AST : sig ... end

module Transform : sig ... end

module CNF : sig ... end

module CYK : sig ... end

val key : string

AST

include module type of struct include AST end

Context Free Grammars

module NonTerminal = AST.NonTerminal

type nonterminal = NonTerminal.t

type terminal = Common.Letter.t

type cfg = AST.cfg

Other Types

type component = AST.component =

| T of terminal

| N of nonterminal

type production = component list

Reading

val entrypoints : cfg -> nonterminal list

val productions : ?from_:nonterminal -> cfg -> (nonterminal * production) Stdlib.Seq.t

val productions_list : ?from_:nonterminal -> cfg -> (nonterminal * production) list

Same as productions but returns a list.

val nonterminals : cfg -> nonterminal Stdlib.Seq.t

val nonterminals_list : cfg -> nonterminal list

val terminals : cfg -> terminal Stdlib.Seq.t

val terminals_list : cfg -> terminal list

val alphabet : cfg -> Common.Alphabet.t

Creating

val empty_cfg : cfg

val fresh_nonterminal : ?hint:string -> unit -> nonterminal

val add_entrypoint : nonterminal -> cfg -> cfg

val add_entrypoints : nonterminal list -> cfg -> cfg

val add_production : nonterminal -> production -> cfg -> cfg

val add_productions : nonterminal -> production list -> cfg -> cfg

val add_productionss : (nonterminal * production) list -> cfg -> cfg

val update_productions : (nonterminal -> production -> (nonterminal * production) list) -> cfg -> cfg

type t = AST.cfg

CST Parsing & Printing

module ParsingFunctions : sig ... end

val cst_from_string : ?filename:string -> string -> (CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located, Stdlib.Lexing.position * Stdlib.Lexing.position * string) Stdlib.result

val cst_from_channel : ?filename:string -> Stdlib.in_channel -> (CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located, Stdlib.Lexing.position * Stdlib.Lexing.position * string) Stdlib.result

val cst_from_file : string -> (CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located, Stdlib.Lexing.position * Stdlib.Lexing.position * string) Stdlib.result

val cst_from_string_exn : ?filename:string -> string -> CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located

val cst_from_channel_exn : ?filename:string -> Stdlib.in_channel -> CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located

val cst_from_file_exn : string -> CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located

val pp_cst : Stdlib.Format.formatter -> CFG__.CST.rule list Common.CSTHelpers.located -> unit

val cst_to_string : CFG__.CST.rule list Common.CSTHelpers.located -> string

val cst_to_channel : Stdlib.out_channel -> CFG__.CST.rule list Common.CSTHelpers.located -> unit

val cst_to_file : string -> CFG__.CST.rule list Common.CSTHelpers.located -> unit

val cst_to_ast : CFG__.CST.rule Common.CSTHelpers.located list Common.CSTHelpers.located -> AST.cfg

val ast_to_cst : AST.cfg -> CFG__.CST.rule list Common.CSTHelpers.located

AST Parsing & Printing

val from_channel_exn : Stdlib.in_channel -> AST.cfg

val from_string_exn : string -> AST.cfg

val from_file_exn : string -> AST.cfg

val pp : Stdlib.Format.formatter -> AST.cfg -> unit

val to_channel : Stdlib.out_channel -> AST.cfg -> unit

val to_string : AST.cfg -> string

val to_file : string -> AST.cfg -> unit

Words Acceptance

val accepts : AST.cfg -> Common.Word.t -> [> `False | `True ]

Test of acceptance of a word by a CFG. Note that this function works in two steps: given the CFG, it computes its CNF, stores it efficiently and then returns a function which, given a word, checks whether it is accepted. This means that for repetitive uses of the same function, it is much more efficient to store the result of accepts cfg and then call this function rather than call accepts cfg word again on every word.

module IncrementalAcceptance : sig ... end