Module CFG.CYK

Cocke-Younger-Kasami Algorithm

Algorithm to check whether a word is recognised by an arbitrary context free grammar. It can be modified to also provide a parse tree, which is here the case. It runs in O(n²⋅r) with n the size of the word and r the size of the grammar.

type parsetree =

| EProd

| TProd of AST.terminal

| NTProd of CNF.nonterminal * parsetree * CNF.nonterminal * parsetree

val parse : CNF.t -> Common.Word.t -> parsetree option

Try to parse the given word with the given grammar and to return a parse tree. Returns None if the word is not recognised.

val accepts : CNF.t -> Common.Word.t -> bool

Checks whether a given grammar recognises the given word.

Incremental Version

Incremental version of the CYK algorithm. When it is used to parse a whole word, it runs in O(n²⋅r⋅log(max n r)) with n the size of the word and r the size of the grammar. The logarithm comes from the use of maps instead of arrays in the implementation.

module State : sig ... end