Module CFG.Transform

Various Transformations of CFG

val extract_entrypoint : AST.cfg -> AST.cfg

Eliminate the entrypoint from right-hand sides. Returns a CFG recognising the same language but with only one entrypoint that does not appear on the right-hand side of rules.

val eliminate_terminals : AST.cfg -> AST.cfg

Eliminate terminals from right-hand sides, except if there is only one terminal and non nonterminal. Returns a CFG recognising the same language but with rules producing either only nonterminals or only one terminal.

val eliminate_terminals_after_nonterminal : AST.cfg -> AST.cfg

Same as eliminate_terminals except only at the right of nonterminals.

val limit_to_two_nonterminals : AST.cfg -> AST.cfg

Eliminate right-hand sides with more than two nonterminals. Returns a CFG recognising the same language but with rules producing at most two nonterminals.

val inline_epsilon_productions : AST.cfg -> AST.cfg

Eliminate ε-productions. Returns a CFG recognising the same language but with no ε-production, except possibly on the entrypoints.

val merge_unit_cycles : AST.cfg -> AST.cfg

Merge unit cycles, that is cycles formed by productions of the form A -> B. When such a cycle occurs, the nonterminals all recognise the same language and can be merged together. Returns a CFG recognising the same language but without such cycles.

val inline_unit_productions : AST.cfg -> AST.cfg

Inline unit productions, that is productions of the form A -> B. We can replace such productions by inline the productions of B inside A. Returns a CFG recognising the same language but without such productions.

val remove_unreachable : AST.cfg -> AST.cfg

Remove unreachable nonterminals for the grammar.

Chomsky Normal Form

Following the presentation of Lange & Leiß 2009.

val start : AST.cfg -> AST.cfg

Alias for extract_entrypoint.

val term : AST.cfg -> AST.cfg

Alias for eliminate_terminals.

val bin : AST.cfg -> AST.cfg

Alias for limit_to_two_nonterminals.

val del : AST.cfg -> AST.cfg

Alias for inline_epsilon_productions.

val unit : AST.cfg -> AST.cfg

Alias for inline_unit_productions.

val chomsky_normal_form : AST.cfg -> AST.cfg

Successive application of start, term, bin, del and unit. Returns a CFG recognising the same language but in Chomsky Normal Form, that is with only one entrypoint and with only rules of the form S -> ε, A -> BC or A -> a where S is the entrypoint, A, B, C are nonterminals (B and C cannot be the entrypoint) and a is a terminal.

val cnf : AST.cfg -> AST.cfg

Alias for chomsky_normal_form.