Module Common.CSTHelpers

type position

type 'a located

val pp_located : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'a located -> Ppx_deriving_runtime.unit

val show_located : (Ppx_deriving_runtime.Format.formatter -> 'a -> Ppx_deriving_runtime.unit) -> 'a located -> Ppx_deriving_runtime.string

val dummy_lexing_position : Stdlib.Lexing.position

val dummy_position : position

val with_position : position -> 'a -> 'a located

val with_positions : Stdlib.Lexing.position -> Stdlib.Lexing.position -> 'a -> 'a located

val pp_lexing_position : Stdlib.Format.formatter -> Stdlib.Lexing.position -> unit

val pp_position : Stdlib.Format.formatter -> position -> unit

val dummily : 'a -> 'a located

val map_add_dummy : ('a -> 'b) -> 'a -> 'b located

val value : 'a located -> 'a

val map_located : ('a -> 'b) -> 'a located -> 'b located

val map_ignore_located : ('a -> 'b) -> 'a located -> 'b

val pp_ignore_located : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a located -> unit

When pp is a printer for values of type 'a, pp_ignore_located pp is the printer for values of type 'a located that ignores locations and simply call pp on the value.

exception SyntaxError of position * string

val syntax_error : ?start_pos:Stdlib.Lexing.position -> ?end_pos:Stdlib.Lexing.position -> ?pos:position -> ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a

val catch_syntax_error : ('a -> 'b) -> 'a -> 'b