Module Ext__ExtSeq

include module type of Stdlib.Seq

type !'a t = unit -> 'a node

and !'a node = 'a Stdlib__seq.node =

| Nil

| Cons of 'a * 'a t

val empty : 'a t

val return : 'a -> 'a t

val cons : 'a -> 'a t -> 'a t

val append : 'a t -> 'a t -> 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

val filter : ('a -> bool) -> 'a t -> 'a t

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

val flat_map : ('a -> 'b t) -> 'a t -> 'b t

val fold_left : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

val iter : ('a -> unit) -> 'a t -> unit

val unfold : ('b -> ('a * 'b) option) -> 'b -> 'a t

val append : 'a t -> 'a t -> 'a t

val flatten : 'a t t -> 'a t

val iter2 : ('a -> 'b -> unit) -> 'a t -> 'b t -> unit

Same as List.iter2 but for sequences. The Invalid_argument exception is raised only when the end is reached.

val map2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

Same as List.map2 but for sequences. The Invalid_argument exception is raised only when the end is reached.