Module LanguagePrefix

type kind =

| Complete

| Incomplete

type t

val from_word_list : kind:kind -> Common.Word.t list -> t

val kind : t -> kind

Kind of the prefix, complete or incomplete. This specifies whether there are more words than the known ones in the language.

val is_complete : t -> bool

val is_incomplete : t -> bool

val words : t -> Common.Word.t Stdlib.Seq.t

Sequence of words in the prefix. The sequence is finite. The words are given in sorted order with respect to Word.compare.

val words_list : t -> Common.Word.t list

Same as words but the result is given as a list.

val from_file_exn : string -> LanguagePrefix__.AST.t

val from_channel_exn : Stdlib.in_channel -> LanguagePrefix__.AST.t

val from_string_exn : string -> LanguagePrefix__.AST.t

val to_string : 'a -> 'b

val to_file : 'a -> 'b -> 'c

val to_channel : 'a -> 'b -> 'c

val length : t -> int

val alphabet : t -> Common.Alphabet.t

val accepts : t -> Common.Word.t -> [> `DontKnow of [> `SizeLimit ] | `False | `True ]

module IncrementalAcceptance : sig ... end