Module CFG.CNF

Chomsky Normal Form

Efficient representation of grammars in Chomsky normal forms.

A context-free grammar is said to be in Chomsky normal form if all of its production rules are of the form: A -> BC or A -> a or S -> ε where A, B and C are nonterminal symbols, the letter a is a terminal symbol, S is the start symbol and ε denotes the empty string. Also, neither B nor C may be the start symbol, and the third production rule can only appear if ε is in the language produced by the grammar.

type cnf = {

empty : bool;

t_prod : AST.terminal list array;

nt_prod : (int * int) list array;

hints : AST.nonterminal array;

}

val from_cfg : AST.cfg -> cnf

type parsetree =

| EProd

| TProd of AST.terminal

| NTProd of int * parsetree * int * parsetree

val parse : cnf -> AST.terminal array -> parsetree option

val parse : cnf -> Common.Word.t -> parsetree option

val pp_parsetree : cnf -> Stdlib.Format.formatter -> (int * parsetree) -> unit

val accepts : cnf -> Common.Word.t -> bool